requirements.txt
python-dotenv
requests
boto3
SQLAlchemy

